🧑‍💻 Move queue things to docker compose
Since the official docker build of Arena doesn't support Apple silicon, we needed to build it ourselves. This seemed like a good time to move that stuff into docker compose anyway, to reduce the number of commands necessary to get the platform up, and make all those settings more visible.